Extracted upload spec from SampleDataSpec.
For data uploads, we will need a spec object that defines how columns/fields
work, but not where to read the data from or where to store it.
